input: misc: fpc1145: Fix and optimize regulators handling
It's useless to request and deregister a regulator everytime
we bring one up/down... also, it's expensive to search for
a regulator through its name in a struct containing a string
to get a match.

Optimize the regulators handling by introducing an enum
containing the regulator map, request the regulators only
at probe time, differentiate between required and optional
rails, and use devm functions to benefit from the auto put
on device detach.
This change also fixes the kernel logspam because of
optional regulators not found.

Change-Id: I13083e9e22f97d29f89ad662eeed31effd244686
Signed-off-by: Chippa-a <vusal1372@gmail.com>